+++ src/Krystal/Db/Sql/QueryBuilderInterface.php
@@ -383,6 +383,36 @@ interface QueryBuilderInterface
     public function orWhere($key, $operator, $value, $filter = false);
 
     /**
+     * An alternative of ORDER BY RAND() for the first WHERE condition
+     * 
+     * @param string $column Column to be sorted by in random order
+     * @param string $table Target table name
+     * @throws \RuntimeException if no table is selected and no table name provided 
+     * @return \Krystal\Db\Sql\QueryBuilder
+     */
+    public function whereRandom($column, $table = null);
+
+    /**
+     * An alternative of ORDER BY RAND() for the second WHERE condition
+     * 
+     * @param string $column Column to be sorted by in random order
+     * @param string $table Target table name
+     * @throws \RuntimeException if no table is selected and no table name provided 
+     * @return \Krystal\Db\Sql\QueryBuilder
+     */
+    public function andWhereRandom($column, $table = null);
+
+    /**
+     * An alternative of ORDER BY RAND() for the second WHERE ... OR condition
+     * 
+     * @param string $column Column to be sorted by in random order
+     * @param string $table Target table name
+     * @throws \RuntimeException if no table is selected and no table name provided 
+     * @return \Krystal\Db\Sql\QueryBuilder
+     */
+    public function orWhereRandom($column, $table = null);
+
+    /**
      * Appends OR WHERE with != operator
      * 
      * @param string $column